Adding some remote dependencies to the capfile.
The capfile now restarts the queue_fetcher daemon for a deployment

Removed some cruft from the Capfile. And probably added some.

Added an OnHungupCall hook which tries to log the agent out when they hang up.

Changed some timeouts to be much more reasonable.

Dialing an extension will not go to voicemail when they hangup if the dial was success
ful.